move config to utils/
