--- src/lib/components/Treebank/config/arethusa-configs
@@ -1,1 +1,0 @@
-../../../../../vendor/arethusa-configs/dist/
