+++ src/lib/utils/config/arethusa-configs
@@ -1,0 +1,1 @@
+../../../../vendor/arethusa-configs/dist/
